requirements.txt
numpy
langchain
langchain-community
langchain-core
langchain-huggingface
faiss-cpu
gitpython
sentence-transformers
accelerate